README Multiple IRQ
The system awaits for an Interrupt from remote sensor(s) and once the sensor is triggered it sends the action signal to the MCU. The MCU then stops everything it is doing at the moment and issues a statement (among which an LED at RD0 and RC0) then continues from the PC counter it stored prior to the interrupt. There are two interrupts in the Assembly code, no priority was set since interrupts are based in "Interrupt-On-Change" occurring on PORT B pins. The Sensor trigger signals are sent via a custom protocol I designed for a product I am working on. This is the base line software of my product.

Feel free to use it in your own design.